requirements.txt
pytube
customtkinter
Pillow
urllib
ffmpeg-python